(BSR)[API] fix: remove repository/user_queries.py
